requirements.txt
rich
openai
python-dotenv